+++ Commons/TransactionApi/src/main/java/com/mfino/transactionapi/handlers/mobileshopping/impl/PurchaseInquiryHandlerImpl.java
@@ -220,7 +220,7 @@ public class PurchaseInquiryHandlerImpl extends FIXMessageHandler implements Pur
 		log.info("Got the response from backend .The notification code is : "+transactionResponse.getCode()+" and the result: "+transactionResponse.isResult());
 
 		if (transactionResponse.getTransactionId() !=null) {
-			sctl.setTransactionid(BigDecimal.valueOf(transactionResponse.getTransactionId()));
+			sctl.setTransactionid(transactionResponse.getTransactionId());
 			purchaseInquiry.setTransactionID(transactionResponse.getTransactionId());
 			result.setTransactionID(transactionResponse.getTransactionId());
 			transactionChargingService.saveServiceTransactionLog(sctl);
