+++ Commons/TransactionApi/src/main/java/com/mfino/transactionapi/handlers/payment/impl/BSIMBillInquiryHandlerImpl.java
@@ -262,8 +262,8 @@ public class BSIMBillInquiryHandlerImpl extends FIXMessageHandler implements BSI
         
 		  sctl = sctlService.getBySCTLID(sctlId);
 		if (transactionResponse.getTransactionId()!=null) {
-			sctl.setTransactionid(BigDecimal.valueOf(transactionResponse.getTransactionId()));
-			sctl.setCommoditytransferid(BigDecimal.valueOf(transactionResponse.getTransferId()));
+			sctl.setTransactionid(transactionResponse.getTransactionId());
+			sctl.setCommoditytransferid(transactionResponse.getTransferId());
 			billInquiry.setTransactionID(transactionResponse.getTransactionId());
 			result.setTransactionID(transactionResponse.getTransactionId());
 			transactionChargingService.saveServiceTransactionLog(sctl);
